+++ config/webpack/environment.js
@@ -2,7 +2,7 @@ const { environment } = require('@rails/webpacker')
 const typescript =  require('./loaders/typescript')
 
 const webpack = require('webpack')
-environment.plugins.append('Provide', new webpack.ProvidePlugin({
+environment.plugins.prepend('Provide', new webpack.ProvidePlugin({
   $: 'jquery',
   jQuery: 'jquery',
   Popper: ['popper.js', 'default']
